perf(dsp): compute CMR and whitening in float32
The streaming chain now carries slabs and intermediates as float32 and
applies the common-median-reference and ZCA whitening matmul in float32,
matching SpikeInterface's float32 processing. The bandpass filter still
runs its per-channel math in float64 (it is ~1/40th of the runtime, so its
accuracy is kept), but its output, CMR, and whitening are float32.

This halves the in-flight memory per chunk and roughly halves the cost of
the whitening GEMM, which dominates the chain at Neuropixels channel counts.
The whitening matrix is still estimated in float64 (cov + symmetric
eigendecomposition) for accuracy, then stored as float32 for application.

Correctness vs the float64 scipy reference now sits at the float32 floor
(~1e-5 relative, well within the tolerances); 24 Rust + 21 Python tests pass.

+++ src/dsp/cmr.rs
@@ -1,6 +1,6 @@
 use numpy::ndarray::Array2;
 
-fn median_of(buf: &mut [f64]) -> f64 {
+fn median_of(buf: &mut [f32]) -> f32 {
     let n = buf.len();
     let mid = n / 2;
     buf.select_nth_unstable_by(mid, |a, b| a.partial_cmp(b).unwrap());
@@ -8,12 +8,12 @@ fn median_of(buf: &mut [f64]) -> f64 {
         buf[mid]
     } else {
         let upper = buf[mid];
-        let lower = buf[..mid].iter().copied().fold(f64::NEG_INFINITY, f64::max);
+        let lower = buf[..mid].iter().copied().fold(f32::NEG_INFINITY, f32::max);
         (lower + upper) / 2.0
     }
 }
 
-pub fn common_median_reference(chunk: &mut Array2<f64>, scratch: &mut Vec<f64>) {
+pub fn common_median_reference(chunk: &mut Array2<f32>, scratch: &mut Vec<f32>) {
     for mut row in chunk.rows_mut() {
         scratch.clear();
         scratch.extend(row.iter().copied());
@@ -31,17 +31,17 @@ mod tests {
 
     #[test]
     fn subtracts_per_sample_median_odd_channels() {
-        let mut chunk = array![[1.0, 2.0, 3.0], [10.0, 40.0, 70.0]];
+        let mut chunk = array![[1.0f32, 2.0, 3.0], [10.0, 40.0, 70.0]];
         let mut scratch = Vec::new();
         common_median_reference(&mut chunk, &mut scratch);
-        assert_eq!(chunk, array![[-1.0, 0.0, 1.0], [-30.0, 0.0, 30.0]]);
+        assert_eq!(chunk, array![[-1.0f32, 0.0, 1.0], [-30.0, 0.0, 30.0]]);
     }
 
     #[test]
     fn even_channels_average_two_middles() {
-        let mut chunk = array![[1.0, 2.0, 3.0, 4.0]];
+        let mut chunk = array![[1.0f32, 2.0, 3.0, 4.0]];
         let mut scratch = Vec::new();
         common_median_reference(&mut chunk, &mut scratch);
-        assert_eq!(chunk, array![[-1.5, -0.5, 0.5, 1.5]]);
+        assert_eq!(chunk, array![[-1.5f32, -0.5, 0.5, 1.5]]);
     }
 }
